% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Search-Methods.R
\name{move_MoveObject}
\alias{move_MoveObject}
\title{Query a \code{\link{MoveObject}}}
\usage{
move_MoveObject(ix, dest)
}
\arguments{
\item{ix}{integer indicating index of current \code{\link{Site}}}

\item{dest}{character in {f,l,m,s} indicating destination resource}
}
\value{
list with elements \code{newIx} and \code{dist}
}
\description{
Given an input index and destination resource, return the index of the \code{\link{Site}}
that the mosquito moves to, and distance traveled. Movement within a site (peri-domestic movement) is
accounted for through weights to stay 'here'.
\itemize{
\item this method is bound to \code{MoveObject$move}
}
}
