% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Search-Class.R
\docType{data}
\name{move_obj_Kernel}
\alias{move_obj_Kernel}
\title{Movement Probabilities for 'Kernel'}
\format{An object of class \code{list} of length 13.}
\usage{
move_obj_Kernel
}
\description{
Gives relevant movement information called in \code{\link{move_MoveObject}} for
a single \code{\link{Site}}.
\itemize{
\item here: peri-domestic movement (multiple resources found at same site)
\item near: movement within the same tile
\item far: movement between tiles
}
}
\keyword{datasets}
