% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Search-Class.R
\docType{class}
\name{MoveObject}
\alias{MoveObject}
\title{MoveObject Abstract Base Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
MoveObject
}
\description{
A movement object lives in a \code{\link{Tile}}
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item move(ix, dest): this function must return a (possibly the same) site index and distance flown to calculate wing tattering
\itemize{
\item ix: index of \code{\link{Site}} the mosquito is currently at
\item dest: character in {f,l,m,s} indicating the resource the mosquito is trying to get to
}
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
