% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Auxiliary.R
\name{sample_aux}
\alias{sample_aux}
\title{Sample Indices}
\usage{
sample_aux(x, ...)
}
\description{
Wrapper for \code{\link{sample}} that will not lead to unexpected behavior when \code{x} is of length 1.
}
