% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Class.R
\docType{class}
\name{Site}
\alias{Site}
\title{Site Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Site
}
\description{
A site is a location where resources can be found.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
